用户注册功能实现

+++ src/main/java/com/miaoshaproject/service/model/UserModel.java
@@ -44,7 +44,7 @@ public class UserModel {
         return gender;
     }
 
-    public void setGender(Byte gender) {
+    public void setGender(Integer gender) {
         this.gender = gender;
     }
